# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RMQTT Things API",
    "description": "RMQTT IoT Things Management API",
    "license": {
      "name": ""
    },
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/access/auth": {
      "post": {
        "tags": ["access"],
        "operationId": "auth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "allow or deny",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/alarm": {
      "get": {
        "tags": ["admin"],
        "operationId": "list_alarms",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "description": "Filter by product ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "description": "Filter by device ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Filter by level: info / warning / critical",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "acknowledged",
            "in": "query",
            "description": "Filter by acknowledged status",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status: active / acknowledged / cleared",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number, default 1",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Page size, default 10",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alarm record list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/admin/alarm-rule": {
      "get": {
        "tags": ["admin"],
        "operationId": "list_alarm_rules",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "description": "Filter by product ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "description": "Filter by enabled status",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number, default 1",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Page size, default 10",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alarm rule list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      },
      "post": {
        "tags": ["admin"],
        "operationId": "create_alarm_rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlarmRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Alarm rule created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmRuleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid product_id or request body"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
    "/api/admin/ca/cert": {
      "get": {
        "tags": ["admin"],
        "operationId": "list_certs_handler",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "description": "产品ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "description": "设备ID，可选",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "设备状态: online, offline",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DeviceConnectionStatus"
            }
          },
          {
            "name": "registration_source",
            "in": "query",
            "description": "注册来源: 0=auto, 1=manual",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RegistrationSource"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码，默认为1",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页大小，默认为10",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Certificate list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimplePaginatedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      },
      "post": {
        "tags": ["admin"],
        "operationId": "issue_cert_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueCertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Certificate issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueCertResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
    "/api/admin/device/status": {
      "get": {
        "tags": ["admin"],
        "operationId": "get_device_status",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "description": "产品ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "description": "设备ID，可选",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "设备状态: online, offline",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DeviceConnectionStatus"
            }
          },
          {
            "name": "registration_source",
            "in": "query",
            "description": "注册来源: 0=auto, 1=manual",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RegistrationSource"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码，默认为1",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页大小，默认为10",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device statuses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
    "/api/admin/event": {
      "get": {
        "tags": ["admin"],
        "operationId": "get_event_history",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "description": "产品ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "description": "设备ID，可选",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码，默认为1",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页大小，默认为10",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimplePaginatedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
    "/api/admin/file/upload": {
      "post": {
        "tags": ["admin"],
        "operationId": "admin_file_upload_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Upload policy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
    "/api/admin/property": {
      "get": {
        "tags": ["admin"],
        "operationId": "get_property_latest",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "description": "产品ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "description": "设备ID，可选",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码，默认为1",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页大小，默认为10",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest properties",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimplePaginatedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
    "/api/device/connect": {
      "post": {
        "tags": ["device"],
        "operationId": "device_connect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Device connection stored"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlarmRuleResponse": {
        "type": "object",
        "description": "Single alarm rule response wrapper",
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AlarmRule"
          }
        }
      },
...
      "AuthPayload": {
        "type": "object",
        "required": ["client_id", "protocol", "ipaddress"],
        "properties": {
          "client_id": {
            "type": "string"
          },
          "ipaddress": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "protocol": {},
          "username": {
            "type": ["string", "null"]
          }
        }
      },
...
      "DeviceConnectionStatus": {
        "type": "string",
        "enum": ["Offline", "Online"]
      },
...
      "PaginatedResponse": {
        "type": "object",
        "required": ["data", "pagination"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiAlarmRecord"
            },
            "description": "数据列表"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo",
            "description": "分页信息"
          }
        }
      },
...
      "RegistrationSource": {
        "type": "string",
        "enum": ["Auto", "Manual"]
      },
...
      "SimplePaginatedResponse": {
        "type": "object",
        "required": ["data", "pagination"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CertIssue"
            },
            "description": "数据列表"
          },
          "pagination": {
            "$ref": "#/components/schemas/SimplePaginationInfo",
            "description": "分页信息"
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "X-Auth"
      }
    }
  },
...
}